moves graph components into components folder
