+++ src/components/map-components/Map.scss
@@ -58,7 +58,10 @@
             font-size:1.1rem;
         }
 
-        
+        .chartjs-render-monitor {
+            width:80%;
+            margin: 0 auto;
+        }
 
         .data-nav-top {
             display:flex;
